+++ AlgoritmoRSA/RSAEncoder/Models/RSANumbers.cs
@@ -2,11 +2,11 @@
 {
     public class RSANumbers
     {
-        public int P { get; set; }
-        public int Q { get; set; }
-        public int N { get; set; }
-        public int M { get; set; }
-        public int E { get; set; }
-        public int D { get; set; }
+        public long P { get; set; }
+        public long Q { get; set; }
+        public long N { get; set; }
+        public long M { get; set; }
+        public long E { get; set; }
+        public long D { get; set; }
     }
 }
